pyproject.toml
[project]
name = "rag-github-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.8",
    "gitpython>=3.1.46",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-openai>=1.1.14",
    "langchain-text-splitters>=1.1.2",
    "python-dotenv>=1.2.2",
]
